requirements.txt
arrow==0.14.7
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
ics==0.7
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
TatSu==5.6.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
